Split subcommands to separate files
